Warning fix for old Linux 2.6 for i386
av_beacon_alloc should be unsigned long, not unsigned int to match
signatures of bitmask manupulation functions on i386.


git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@3856 0192ed92-7a03-0410-a25b-9323aeb14dbd

+++ ath/if_athvar.h
@@ -537,7 +537,7 @@ struct ath_vap {
 	atomic_t av_beacon_alloc;       /* set to 1 when the next beacon needs
 					   to be recomputed */
 #else
-	unsigned int av_beacon_alloc;
+	unsigned long av_beacon_alloc;
 #endif
 };
 #define	ATH_VAP(_v)	((struct ath_vap *)(_v))
